* Refactoring of package structures
