+++ notification/src/main/java/ch/unisg/airqueue/notification/commands/NotificationCommand.java
@@ -1,4 +1,4 @@
-package ch.unisg.airqueue.notification.messages;
+package ch.unisg.airqueue.notification.commands;
 
 public class NotificationCommand {
     private String content;
