+++ notification/src/main/java/ch/unisg/airqueue/notification/events/NotificationSentEvent.java
@@ -1,4 +1,4 @@
-package ch.unisg.airqueue.notification.messages;
+package ch.unisg.airqueue.notification.events;
 
 public class NotificationSentEvent {
