+++ payment-mono/src/main/java/ch/unisg/airqueue/payment/commands/GetPaymentCommand.java
@@ -1,4 +1,4 @@
-package ch.unisg.airqueue.payment.messages;
+package ch.unisg.airqueue.payment.commands;
 
 public class GetPaymentCommand {
